serach
The search file allows the user to search for a show in the watched table by entering any value, even 
if there is more than one show, according to the user's input. The user can choose the value they want to 
search for. The show/s is/are displayed in a table format like the listing file.